0318 run with world model training

+++ dreamer/tools.py
@@ -7,6 +7,7 @@ import pathlib
 import re
 import time
 import random
+import pdb
 
 import numpy as np
 
@@ -486,9 +487,9 @@ class DiscDist:
         below = torch.clip(below, 0, len(self.buckets) - 1)
         above = torch.clip(above, 0, len(self.buckets) - 1)
         equal = below == above
-
-        dist_to_below = torch.where(equal, 1, torch.abs(self.buckets[below] - x))
-        dist_to_above = torch.where(equal, 1, torch.abs(self.buckets[above] - x))
+        # pdb.set_trace()
+        dist_to_below = torch.where(equal, 1.0, torch.abs(self.buckets[below] - x.double()))
+        dist_to_above = torch.where(equal, 1.0, torch.abs(self.buckets[above] - x.double()))
         total = dist_to_below + dist_to_above
         weight_below = dist_to_above / total
         weight_above = dist_to_below / total
